+++ src/Microsoft.Data.SqlClient/tests/ManualTests/SQL/RetryLogic/SqlConfigurationManagerReliabilityTest.cs
@@ -45,6 +45,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
             RetryLogicConfigHelper.AssessProvider(cmdProvider, cmdCfg, switchValue);
 
             // check the retry in action
+            RetryLogicTestHelper.CleanRetryEnabledCache();
             s_connectionCRLTest.ConnectionRetryOpenInvalidCatalogFailed(TcpCnnString, cnnProvider);
             s_commandCRLTest.RetryExecuteFail(TcpCnnString, cmdProvider);
             s_commandCRLTest.RetryExecuteUnauthorizedSqlStatementDML(TcpCnnString, cmdProvider);
@@ -65,6 +66,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
             RetryLogicConfigHelper.AssessProvider(cnnProvider, cnnCfg, switchValue);
             RetryLogicConfigHelper.AssessProvider(cmdProvider, cmdCfg, switchValue);
 
+            RetryLogicTestHelper.CleanRetryEnabledCache();
             s_connectionCRLTest.DefaultOpenWithoutRetry(TcpCnnString, cnnProvider);
             s_commandCRLTest.NoneRetriableExecuteFail(TcpCnnString, cmdProvider);
         }
@@ -95,6 +97,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
             object loaderObj = RetryLogicConfigHelper.ReturnLoaderAndProviders(cnnCfg, cmdCfg, switchValue, out SqlRetryLogicBaseProvider cnnProvider, out SqlRetryLogicBaseProvider cmdProvider);
             Assert.NotNull(loaderObj);
 
+            RetryLogicTestHelper.CleanRetryEnabledCache();
             TestConnection(cnnProvider, cnnCfg);
             TestCommandExecute(cmdProvider, cmdCfg);
             TestCommandExecuteAsync(cmdProvider, cmdCfg).Wait();
@@ -127,6 +130,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
             object loaderObj = RetryLogicConfigHelper.ReturnLoaderAndProviders(cnnCfg, cmdCfg, switchValue, out SqlRetryLogicBaseProvider cnnProvider, out SqlRetryLogicBaseProvider cmdProvider);
             Assert.NotNull(loaderObj);
 
+            RetryLogicTestHelper.CleanRetryEnabledCache();
             s_connectionCRLTest.DefaultOpenWithoutRetry(TcpCnnString, cnnProvider);
             s_commandCRLTest.NoneRetriableExecuteFail(TcpCnnString, cmdProvider);
         }
@@ -148,6 +152,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
             Assert.NotNull(loaderObj);
 
             // none retriable logic applies.
+            RetryLogicTestHelper.CleanRetryEnabledCache();
             s_connectionCRLTest.DefaultOpenWithoutRetry(TcpCnnString, cnnProvider);
             s_commandCRLTest.NoneRetriableExecuteFail(TcpCnnString, cmdProvider);
         }
@@ -174,6 +179,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
             RetryLogicConfigHelper.AssessProvider(cmdProvider, cmdCfg, switchValue);
 
             // internal type used to resolve the specified method
+            RetryLogicTestHelper.CleanRetryEnabledCache();
             s_connectionCRLTest.ConnectionRetryOpenInvalidCatalogFailed(TcpCnnString, cnnProvider);
             s_commandCRLTest.RetryExecuteFail(TcpCnnString, cmdProvider);
         }
